apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
psycopg[binary]==3.2.1
duckdb==1.1.3
databricks-sdk==0.36.0
databricks-sql-connector==3.5.0
pytest==8.3.2
